# openapi/crm.json.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "F-Grade Corporate AI Backend",
    "description": "REST API for Zoho CRM, Books, Inventory, People, and Analytics through the F-Grade Corporate AI Backend.",
    "version": "1.0.0",
    "contact": {
      "name": "F-Grade"
    },
    "license": {
      "name": "Internal Use Only"
    }
  },
  "servers": [
    {
      "url":  "https://unweighted-elois-trinomially.ngrok-free.dev",
      "description": "Local Development"
    }
  ],
  "tags": [
    {
      "name": "CRM",
      "description": "Zoho CRM endpoints"
    }
  ],
  "security": [],
  "paths": {
    "/api/crm/leads": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Leads",
        "description": "Returns CRM lead records from Zoho CRM.",
        "operationId": "getLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/contacts": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Contacts",
        "description": "Returns CRM contact records from Zoho CRM.",
        "operationId": "getContacts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/accounts": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Accounts",
        "description": "Returns CRM account records from Zoho CRM.",
        "operationId": "getAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/deals": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Deals",
        "description": "Returns CRM deal records from Zoho CRM.",
        "operationId": "getDeals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/tasks": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Tasks",
        "description": "Returns CRM task records from Zoho CRM.",
        "operationId": "getTasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/events": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Events",
        "description": "Returns CRM event records from Zoho CRM.",
        "operationId": "getEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/calls": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Calls",
        "description": "Returns CRM call records from Zoho CRM.",
        "operationId": "getCalls",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/meetings": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Meetings",
        "description": "Returns CRM meeting records from Zoho CRM.",
        "operationId": "getMeetings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/notes": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Notes",
        "description": "Returns CRM notes from Zoho CRM.",
        "operationId": "getNotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/products": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Products",
        "description": "Returns CRM products from Zoho CRM.",
        "operationId": "getProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/vendors": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Vendors",
        "description": "Returns CRM vendors from Zoho CRM.",
        "operationId": "getVendors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/quotes": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Quotes",
        "description": "Returns CRM quotes from Zoho CRM.",
        "operationId": "getQuotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/salesorders": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Sales Orders",
        "description": "Returns CRM sales orders from Zoho CRM.",
        "operationId": "getSalesOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/purchaseorders": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Purchase Orders",
        "description": "Returns CRM purchase orders from Zoho CRM.",
        "operationId": "getPurchaseOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/campaigns": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Campaigns",
        "description": "Returns CRM campaigns from Zoho CRM.",
        "operationId": "getCampaigns",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/cases": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Cases",
        "description": "Returns CRM cases from Zoho CRM.",
        "operationId": "getCases",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/solutions": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Solutions",
        "description": "Returns CRM solutions from Zoho CRM.",
        "operationId": "getSolutions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/users": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Users",
        "description": "Returns CRM users from Zoho CRM.",
        "operationId": "getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/enterprise-leads": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get Enterprise Leads",
        "description": "Returns enterprise lead records from Zoho CRM.",
        "operationId": "getEnterpriseLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/partners": {
  "get": {
    "tags": ["CRM"],
    "summary": "Get CRM Partners",
    "description": "Returns partner records from Zoho CRM.",
    "operationId": "getPartners",
    "parameters": [],
    "responses": {
      "200": {
        "description": "Request succeeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RecordListResponse"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
},

"/api/crm/service-provider": {
  "get": {
    "tags": ["CRM"],
    "summary": "Get CRM Service Providers",
    "description": "Returns service provider records from Zoho CRM.",
    "operationId": "getServiceProviders",
    "parameters": [],
    "responses": {
      "200": {
        "description": "Request succeeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RecordListResponse"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
},

"/api/crm/co-operative-banks": {
  "get": {
    "tags": ["CRM"],
    "summary": "Get CRM Co-operative Banks",
    "description": "Returns co-operative bank records from Zoho CRM.",
    "operationId": "getCoOperativeBanks",
    "parameters": [],
    "responses": {
      "200": {
        "description": "Request succeeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RecordListResponse"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
},

"/api/crm/documents": {
  "get": {
    "tags": ["CRM"],
    "summary": "Get CRM Documents",
    "description": "Returns document records from Zoho CRM.",
    "operationId": "getDocuments",
    "parameters": [],
    "responses": {
      "200": {
        "description": "Request succeeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RecordListResponse"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
},

"/api/crm/projects": {
  "get": {
    "tags": ["CRM"],
    "summary": "Get CRM Projects",
    "description": "Returns project records.",
    "operationId": "getProjects",
    "parameters": [],
    "responses": {
      "200": {
        "description": "Request succeeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RecordListResponse"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
},
    "/api/crm/renewal-accounts": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get Renewal Accounts",
        "description": "Returns renewal account records from Zoho CRM.",
        "operationId": "getRenewalAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crm/org": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get CRM Organization",
        "description": "Returns organization information from Zoho CRM.",
        "operationId": "getOrg",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "service": {
            "type": "string",
            "example": "F-Grade Corporate AI Backend"
          }
        }
      },
      "RecordListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      }
    },
    "responses": {}
  }
}